requirements.txt
Django
djangorestframework
psycopg2-binary
dj-database-url
djangorestframework-jwt
python-dotenv
